+++ packages/spotlight/CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact spotlight module, newest
 
 - `spotlight` to not explicitly `blur()` the currently focused element when focusing another allowing the platform to manage blurring before focus.
 
+### Fixed
+
+- `spotlight` to correctly set focus when the window is activated
+
 ## [2.1.4] - 2018-09-17
 
 ### Fixed