+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `spotlight` to correctly set focus when the window is activated
+
+## [2.1.4] - 2018-09-17
+
+### Fixed
+
+- `spotlight/Spottable` to respect paused state when it becomes enabled
+
+## [2.1.3] - 2018-09-10
+
+No significant changes.
+
 ## [2.1.2] - 2018-09-04
 
 ### Fixed